requirements.txt
Flask==2.1.2
Flask-Limiter==2.4.5.1
pyjwt==1.7.1
bcrypt==3.1.7
